# open-api/openapi.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "v0.6.0",
    "title" : "Aries Cloud Agent"
  },
...
  "security" : [ {
    "AuthorizationHeader" : [ ]
  } ],
  "paths" : {
...
    "/multitenancy/wallet" : {
      "post" : {
        "tags" : [ "multitenancy" ],
        "summary" : "Create a subwallet",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/CreateWalletRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/CreateWalletResponse"
            }
          }
        }
      }
    },
    "/multitenancy/wallet/{wallet_id}" : {
      "get" : {
        "tags" : [ "multitenancy" ],
        "summary" : "Get a single subwallet",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "wallet_id",
          "in" : "path",
          "description" : "Subwallet identifier",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/WalletRecord"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "multitenancy" ],
        "summary" : "Update a subwallet",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UpdateWalletRequest"
          }
        }, {
          "name" : "wallet_id",
          "in" : "path",
          "description" : "Subwallet identifier",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/WalletRecord"
            }
          }
        }
      }
    },
    "/multitenancy/wallet/{wallet_id}/remove" : {
      "post" : {
        "tags" : [ "multitenancy" ],
        "summary" : "Remove a subwallet",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/RemoveWalletRequest"
          }
        }, {
          "name" : "wallet_id",
          "in" : "path",
          "description" : "Subwallet identifier",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/MultitenantModuleResponse"
            }
          }
        }
      }
    },
    "/multitenancy/wallet/{wallet_id}/token" : {
      "post" : {
        "tags" : [ "multitenancy" ],
        "summary" : "Get auth token for a subwallet",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/CreateWalletTokenRequest"
          }
        }, {
          "name" : "wallet_id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/CreateWalletTokenResponse"
            }
          }
        }
      }
    },
    "/multitenancy/wallets" : {
      "get" : {
        "tags" : [ "multitenancy" ],
        "summary" : "Query subwallets",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "wallet_name",
          "in" : "query",
          "description" : "Wallet name",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/WalletList"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "AuthorizationHeader" : {
      "description" : "Bearer token. Be sure to preprend token with 'Bearer '",
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "CreateWalletRequest" : {
      "type" : "object",
      "properties" : {
        "image_url" : {
          "type" : "string",
          "example" : "https://aries.ca/images/sample.png",
          "description" : "Image url for this wallet. This image url is publicized            (self-attested) to other agents as part of forming a connection."
        },
        "key_management_mode" : {
          "type" : "string",
          "example" : "managed",
          "description" : "Key management method to use for this wallet.",
          "enum" : [ "managed" ]
        },
        "label" : {
          "type" : "string",
          "example" : "Alice",
          "description" : "Label for this wallet. This label is publicized            (self-attested) to other agents as part of forming a connection."
        },
        "wallet_dispatch_type" : {
          "type" : "string",
          "example" : "default",
          "description" : "Webhook target dispatch type for this wallet.             default - Dispatch only to webhooks associated with this wallet.             base - Dispatch only to webhooks associated with the base wallet.             both - Dispatch to both webhook targets.",
          "enum" : [ "default", "both", "base" ]
        },
        "wallet_key" : {
          "type" : "string",
          "example" : "MySecretKey123",
          "description" : "Master key used for key derivation."
        },
        "wallet_name" : {
          "type" : "string",
          "example" : "MyNewWallet",
          "description" : "Wallet name"
        },
        "wallet_type" : {
          "type" : "string",
          "example" : "indy",
          "description" : "Type of the wallet to create",
          "enum" : [ "askar", "in_memory", "indy" ]
        },
        "wallet_webhook_urls" : {
          "type" : "array",
          "description" : "List of Webhook URLs associated with this subwallet",
          "items" : {
            "type" : "string",
            "example" : "http://localhost:18022/webhooks",
            "description" : "Optional webhook URL to receive webhook messages"
          }
        }
      }
    },
    "CreateWalletResponse" : {
      "type" : "object",
      "required" : [ "key_management_mode", "wallet_id" ],
      "properties" : {
        "created_at" : {
          "type" : "string",
          "example" : "2021-05-05 12:53:31Z",
          "description" : "Time of record creation",
          "pattern" : "^\\d{4}-\\d\\d-\\d\\d[T ]\\d\\d:\\d\\d(?:\\:(?:\\d\\d(?:\\.\\d{1,6})?))?(?:[+-]\\d\\d:?\\d\\d|Z|)$"
        },
        "key_management_mode" : {
          "type" : "string",
          "description" : "Mode regarding management of wallet key",
          "enum" : [ "managed", "unmanaged" ]
        },
        "settings" : {
          "type" : "object",
          "description" : "Settings for this wallet.",
          "properties" : { }
        },
        "state" : {
          "type" : "string",
          "example" : "active",
          "description" : "Current record state"
        },
        "token" : {
          "type" : "string",
          "example" : "eyJhbGciOiJFZERTQSJ9.eyJhIjogIjAifQ.dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk",
          "description" : "Authorization token to authenticate wallet requests"
        },
        "updated_at" : {
          "type" : "string",
          "example" : "2021-05-05 12:53:31Z",
          "description" : "Time of last record update",
          "pattern" : "^\\d{4}-\\d\\d-\\d\\d[T ]\\d\\d:\\d\\d(?:\\:(?:\\d\\d(?:\\.\\d{1,6})?))?(?:[+-]\\d\\d:?\\d\\d|Z|)$"
        },
        "wallet_id" : {
          "type" : "string",
          "example" : "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "description" : "Wallet record ID"
        }
      }
    },
    "CreateWalletTokenRequest" : {
      "type" : "object",
      "properties" : {
        "wallet_key" : {
          "type" : "string",
          "example" : "MySecretKey123",
          "description" : "Master key used for key derivation. Only required for             unamanged wallets."
        }
      }
    },
    "CreateWalletTokenResponse" : {
      "type" : "object",
      "properties" : {
        "token" : {
          "type" : "string",
          "example" : "eyJhbGciOiJFZERTQSJ9.eyJhIjogIjAifQ.dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk",
          "description" : "Authorization token to authenticate wallet requests"
        }
      }
    },
...
    "MultitenantModuleResponse" : {
      "type" : "object"
    },
...
    "RemoveWalletRequest" : {
      "type" : "object",
      "properties" : {
        "wallet_key" : {
          "type" : "string",
          "example" : "MySecretKey123",
          "description" : "Master key used for key derivation. Only required for             unmanaged wallets."
        }
      }
    },
...
    "UpdateWalletRequest" : {
      "type" : "object",
      "properties" : {
        "image_url" : {
          "type" : "string",
          "example" : "https://aries.ca/images/sample.png",
          "description" : "Image url for this wallet. This image url is publicized            (self-attested) to other agents as part of forming a connection."
        },
        "label" : {
          "type" : "string",
          "example" : "Alice",
          "description" : "Label for this wallet. This label is publicized            (self-attested) to other agents as part of forming a connection."
        },
        "wallet_dispatch_type" : {
          "type" : "string",
          "example" : "default",
          "description" : "Webhook target dispatch type for this wallet.             default - Dispatch only to webhooks associated with this wallet.             base - Dispatch only to webhooks associated with the base wallet.             both - Dispatch to both webhook targets.",
          "enum" : [ "default", "both", "base" ]
        },
        "wallet_webhook_urls" : {
          "type" : "array",
          "description" : "List of Webhook URLs associated with this subwallet",
          "items" : {
            "type" : "string",
            "example" : "http://localhost:18022/webhooks",
            "description" : "Optional webhook URL to receive webhook messages"
          }
        }
      }
    },
...
    "WalletList" : {
      "type" : "object",
      "properties" : {
        "results" : {
          "type" : "array",
          "description" : "List of wallet records",
          "items" : {
            "$ref" : "#/definitions/WalletRecord"
          }
        }
      }
    },
...
    "WalletRecord" : {
      "type" : "object",
      "required" : [ "key_management_mode", "wallet_id" ],
      "properties" : {
        "created_at" : {
          "type" : "string",
          "example" : "2021-05-05 12:53:31Z",
          "description" : "Time of record creation",
          "pattern" : "^\\d{4}-\\d\\d-\\d\\d[T ]\\d\\d:\\d\\d(?:\\:(?:\\d\\d(?:\\.\\d{1,6})?))?(?:[+-]\\d\\d:?\\d\\d|Z|)$"
        },
        "key_management_mode" : {
          "type" : "string",
          "description" : "Mode regarding management of wallet key",
          "enum" : [ "managed", "unmanaged" ]
        },
        "settings" : {
          "type" : "object",
          "description" : "Settings for this wallet.",
          "properties" : { }
        },
        "state" : {
          "type" : "string",
          "example" : "active",
          "description" : "Current record state"
        },
        "updated_at" : {
          "type" : "string",
          "example" : "2021-05-05 12:53:31Z",
          "description" : "Time of last record update",
          "pattern" : "^\\d{4}-\\d\\d-\\d\\d[T ]\\d\\d:\\d\\d(?:\\:(?:\\d\\d(?:\\.\\d{1,6})?))?(?:[+-]\\d\\d:?\\d\\d|Z|)$"
        },
        "wallet_id" : {
          "type" : "string",
          "example" : "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "description" : "Wallet record ID"
        }
      }
    },
...
  }
}